+++ src/js/smooth-scroll.js
@@ -7,7 +7,8 @@
         defaults: {
             duration: 1000,
             transition: 'easeOutExpo',
-            offset: 0
+            offset: 0,
+            complete: function(){}
         },
 
         init: function() {
@@ -27,8 +28,8 @@
                     target = docheight - winheight;
                 }
 
-                // animate to target and set the hash to the window.location after the animation
-                $("html,body").stop().animate({scrollTop: target}, $this.options.duration, $this.options.transition);
+                // animate to target, fire callback when done
+                $("html,body").stop().animate({scrollTop: target}, $this.options.duration, $this.options.transition).promise().done($this.options.complete);
 
                 // cancel default click action
                 return false;